Code/requirements.txt
numpy
nibabel
monai
torchvision
wandb
tqdm
matplotlib